[Platinum V] Title: 버블 소트, Time: 100 ms, Memory: 7172 KB -BaekjoonHub

+++ 백준/Platinum/1517. 버블 소트/버블 소트.cc
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <bits/stdc++.h> 
 using namespace std;
-vector<long long> arr;
-long long temp[500001];
+vector<int> arr;
+int temp[500001];
 long long r = 0;
